requirements.txt
GitPython
tinydb